requirements.txt
requirements.txt
Flask==2.0.1
Flask-Login==0.5.0
pymongo==3.12.0
python-telegram-bot==13.7
stripe==2.60.0
Werkzeug==2.0.1